refactor(web): split RoutineCalendarPanel.tsx — extract hero stats and month grid
